requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
pgvector==0.2.4
numpy==1.24.3
pandas==2.0.3
gunicorn==21.2.0
